--- frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/api/DementiaHomeService.kt
@@ -1,16 +1,0 @@
-package kr.ac.tukorea.whereareu.data.api
-
-import kr.ac.tukorea.whereareu.data.model.ResponseBody
-import kr.ac.tukorea.whereareu.data.model.home.LocationInfo
-import kr.ac.tukorea.whereareu.data.model.home.PostLocationInfoResponse
-import retrofit2.Response
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-interface DementiaHomeService {
-    @POST("receive-location-info")
-    suspend fun postLocationInfo(
-        @Body request: LocationInfo
-    ): Response<PostLocationInfoResponse>
-
-}
