--- frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/repository/home/DementiaHomeRepository.kt
@@ -1,9 +1,0 @@
-package kr.ac.tukorea.whereareu.data.repository.home
-
-import kr.ac.tukorea.whereareu.data.model.home.LocationInfo
-import kr.ac.tukorea.whereareu.data.model.home.PostLocationInfoResponse
-import kr.ac.tukorea.whereareu.util.network.NetworkResult
-
-interface DementiaHomeRepository {
-    suspend fun postLocationInfo(request: LocationInfo): NetworkResult<PostLocationInfoResponse>
-}
