--- frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/repository/home/NokHomeRepository.kt
@@ -1,8 +1,0 @@
-package kr.ac.tukorea.whereareu.data.repository.home
-
-import kr.ac.tukorea.whereareu.data.model.home.GetLocationInfoResponse
-import kr.ac.tukorea.whereareu.util.network.NetworkResult
-
-interface NokHomeRepository {
-    suspend fun getDementiaLocationInfo(dementiaKey: String): NetworkResult<GetLocationInfoResponse>
-}
