+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/presentation/dementia/home/DementiaHomeFragment.kt
@@ -1,0 +1,19 @@
+package kr.ac.tukorea.whereareu.presentation.dementia
+
+import android.content.pm.PackageManager
+import android.os.Build
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import kr.ac.tukorea.whereareu.R
+import kr.ac.tukorea.whereareu.databinding.FragmentHomeBinding
+import kr.ac.tukorea.whereareu.presentation.base.BaseFragment
+
+class DementiaHomeFragment: BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
+    override fun initObserver() {
+
+    }
+
+    override fun initView() {
+
+    }
+}
