+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/presentation/nok/setting/SetUpdateTimeDialogFragment.kt
@@ -1,0 +1,4 @@
+package kr.ac.tukorea.whereareu.presentation.nok.setting
+
+class setUpdateTimeDialogFragment {
+}
